Ensure color property is set on the player when restoring from PGN.
Resolves #17.

The internal game engine resigns based on the player's color. When loading the game from the DB, it was not restoring the color property, so the resign command was silently failing. However, in the integration we assumed it worked and ended the game. The turn hadn't progressed due to the resign command failing and as a result, we would display the current player as the victor, and the outcome as '*', because the state of the game hadn't actually changed.