Add preparation for new oauth token exchange endpoint.
This will properly handle Slack's v2 oauth2. The app distribution URL will change and the configuration of the default redirect URL will change to /slack/oauth/v2 after this is deployed.

Once Slack approves the change, any link on the site that is an "Add to Slack" will need to change to the new v2 endpoint.